+++ g_dfmt_p.c
@@ -31,7 +31,7 @@ THIS SOFTWARE.
 
 #include "gdtoaimp.h"
 
- extern ULong NanDflt_d_D2A[2];
+ extern uint32_t NanDflt_d_D2A[2];
 
  char*
 #ifdef KR_headers
@@ -42,7 +42,7 @@ g_dfmt_p(char *buf, double *d, int ndig, size_t bufsize, int nik)
 {
 	static FPI fpi0 = { 53, 1-1023-53+1, 2046-1023-53+1, 1, 0, Int_max };
 	char *b, *s, *se;
-	ULong bits[2], *L, sign;
+	uint32_t bits[2], *L, sign;
 	int decpt, ex, i, mode;
 #ifdef Honor_FLT_ROUNDS
 #include "gdtoa_fltrnds.h"
@@ -55,7 +55,7 @@ g_dfmt_p(char *buf, double *d, int ndig, size_t bufsize, int nik)
 	if (bufsize < (size_t)(ndig + 10))
 		return 0;
 
-	L = (ULong*)d;
+	L = (uint32_t*)d;
 	sign = L[_0] & 0x80000000L;
 	if ((L[_0] & 0x7ff00000) == 0x7ff00000) {
 		/* Infinity or NaN */
