+++ g_ffmt_p.c
@@ -31,7 +31,7 @@ THIS SOFTWARE.
 
 #include "gdtoaimp.h"
 
- extern ULong NanDflt_f_D2A[1];
+ extern uint32_t NanDflt_f_D2A[1];
 
  char*
 #ifdef KR_headers
@@ -42,7 +42,7 @@ g_ffmt_p(char *buf, float *f, int ndig, size_t bufsize, int nik)
 {
 	static FPI fpi0 = { 24, 1-127-24+1,  254-127-24+1, 1, 0, 6 };
 	char *b, *s, *se;
-	ULong bits[1], *L, sign;
+	uint32_t bits[1], *L, sign;
 	int decpt, ex, i, mode;
 #ifdef Honor_FLT_ROUNDS
 #include "gdtoa_fltrnds.h"
@@ -55,7 +55,7 @@ g_ffmt_p(char *buf, float *f, int ndig, size_t bufsize, int nik)
 	if (bufsize < (size_t)(ndig + 10))
 		return 0;
 
-	L = (ULong*)f;
+	L = (uint32_t*)f;
 	sign = L[0] & 0x80000000L;
 	if ((L[0] & 0x7f800000) == 0x7f800000) {
 		/* Infinity or NaN */
