+++ strtoIQ.c
@@ -39,10 +39,10 @@ strtoIQ(const char *s, char **sp, void *a, void *b)
 #endif
 {
 	static FPI fpi = { 113, 1-16383-113+1, 32766-16383-113+1, 1, SI, 0 /*unused*/ };
-	Long exp[2];
+	int32_t exp[2];
 	Bigint *B[2];
 	int k, rv[2];
-	ULong *L = (ULong *)a, *M = (ULong *)b;
+	uint32_t *L = (uint32_t *)a, *M = (uint32_t *)b;
 #ifdef MULTIPLE_THREADS
 	ThInfo *TI = 0;
 #endif
