+++ strtoIg.c
@@ -33,15 +33,15 @@ THIS SOFTWARE.
 
  int
 #ifdef KR_headers
-strtoIg(s00, se, fpi, exp, B, rvp) const char *s00; char **se; const FPI *fpi; Long *exp; Bigint **B; int *rvp;
+strtoIg(s00, se, fpi, exp, B, rvp) const char *s00; char **se; const FPI *fpi; int32_t *exp; Bigint **B; int *rvp;
 #else
-strtoIg(const char *s00, char **se, const FPI *fpi, Long *exp, Bigint **B, int *rvp)
+strtoIg(const char *s00, char **se, const FPI *fpi, int32_t *exp, Bigint **B, int *rvp)
 #endif
 {
 	Bigint *b, *b1;
 	int i, nb, nw, nw1, rv, rv1, swap;
 	unsigned int nb1, nb11;
-	Long e1;
+	int32_t e1;
 #ifdef MULTIPLE_THREADS
 	ThInfo *TI = 0;
 #endif
